backend/fastapi_app/text-to-slide/requirements.txt
torch==2.4.0
torchvision==0.19.0
torchaudio==2.4.0
diffusers[torch]>=0.30.0
transformers>=4.45.0
accelerate>=0.33.0
python-pptx>=0.6.23
gradio>=4.0.0
sentencepiece